Create User aggregates to handle CRUD

+++ src/main/java/fr/efrei/ordersharingsystem/commands/CreateUserCommand.java
@@ -13,6 +13,6 @@ public record CreateUserCommand(
         @NotBlank @Size(min = 6, max = 255) String password,
         @Size(min = 10, max = 20) String phoneNumber,
         @NotBlank Role role,
-        @NotBlank @Min(value = 1, message = "Minimal parkId is 1") int assignedBowlingParkId
+        @NotBlank @Min(value = 1, message = "Minimal parkId is 1") long assignedBowlingParkId
         ) {
 }
